requirements.txt
python-dotenv
openai
whisper
torch
google-genai
markdown
